+++ mobile/library/common/jni/jni_interface.cc
@@ -168,6 +168,13 @@ Java_io_envoyproxy_envoymobile_engine_JniLibrary_brotliConfigInsert(JNIEnv* env,
 }
 
 extern "C" JNIEXPORT jstring JNICALL
+Java_io_envoyproxy_envoymobile_engine_JniLibrary_persistentDNSCacheConfigInsert(JNIEnv* env,
+                                                                                jclass) {
+  jstring result = env->NewStringUTF(persistent_dns_cache_config_insert);
+  return result;
+}
+
+extern "C" JNIEXPORT jstring JNICALL
 Java_io_envoyproxy_envoymobile_engine_JniLibrary_socketTagConfigInsert(JNIEnv* env, jclass) {
   jstring result = env->NewStringUTF(socket_tag_config_insert);
   return result;