+++ mobile/test/common/integration/base_client_integration_test.h
@@ -35,7 +35,7 @@ std::string defaultConfig();
 //
 // TODO(junr03): move this to derive from the ApiListenerIntegrationTest after moving that class
 // into a test lib.
-class BaseClientIntegrationTest : public BaseIntegrationTest, public Platform::EngineBuilder {
+class BaseClientIntegrationTest : public BaseIntegrationTest {
 public:
   BaseClientIntegrationTest(Network::Address::IpVersion ip_version,
                             const std::string& bootstrap_config = defaultConfig());
@@ -49,6 +49,13 @@ protected:
   void threadRoutine(absl::Notification& engine_running);
   // Must be called manually by subclasses in their TearDown();
   void TearDown();
+  // helpers to access protected functions in the friend class
+  void setOverrideConfigForTests(Platform::EngineBuilder builder, std::string config) {
+    builder.setOverrideConfigForTests(config);
+  }
+  void setAdminAddressPathForTests(Platform::EngineBuilder& builder, std::string admin) {
+    builder.setAdminAddressPathForTests(admin);
+  }
   // Converts TestRequestHeaderMapImpl to Envoy::Platform::RequestHeadersSharedPtr
   Envoy::Platform::RequestHeadersSharedPtr
   envoyToMobileHeaders(const Http::TestRequestHeaderMapImpl& request_headers);
@@ -74,6 +81,7 @@ protected:
   bool override_builder_config_ = false;
   // True if data plane requests are expected in the test; false otherwise.
   bool expect_data_streams_ = true;
+  Platform::EngineBuilder builder_;
 };
 
 } // namespace Envoy