+++ source/common/http/async_client_impl.h
@@ -288,9 +288,7 @@ private:
       return path_match_criterion_;
     }
 
-    const absl::optional<ConnectConfig>& connectConfig() const override {
-      return connect_config_nullopt_;
-    }
+    const ConnectConfigOptRef connectConfig() const override { return connect_config_nullopt_; }
 
     bool includeAttemptCountInRequest() const override { return false; }
     bool includeAttemptCountInResponse() const override { return false; }
@@ -314,7 +312,7 @@ private:
     Router::RouteEntry::UpgradeMap upgrade_map_;
     const std::string& cluster_name_;
     absl::optional<std::chrono::milliseconds> timeout_;
-    static const absl::optional<ConnectConfig> connect_config_nullopt_;
+    static const ConnectConfigOptRef connect_config_nullopt_;
     const std::string route_name_;
     // Pass early data option config through StreamOptions.
     std::unique_ptr<Router::EarlyDataPolicy> early_data_policy_{