+++ test/common/upstream/upstream_impl_test.cc
@@ -3869,6 +3869,30 @@ TEST_F(ClusterInfoImplTest, Timeouts) {
   }
 }
 
+TEST_F(ClusterInfoImplTest, TcpPoolIdleTimeout) {
+  const std::string yaml_base = R"EOF(
+  name: {}
+  type: STRICT_DNS
+  lb_policy: ROUND_ROBIN
+  )EOF";
+
+  const std::string yaml_set_tcp_pool_idle_timeout = yaml_base + R"EOF(
+  typed_extension_protocol_options:
+    envoy.extensions.upstreams.tcp.v3.TcpProtocolOptions:
+      "@type": type.googleapis.com/envoy.extensions.upstreams.tcp.v3.TcpProtocolOptions
+      idle_timeout: {}
+  )EOF";
+
+  auto cluster1 = makeCluster(fmt::format(yaml_base, "cluster1"));
+  EXPECT_EQ(std::chrono::minutes(10), cluster1->info()->tcpPoolIdleTimeout());
+
+  auto cluster2 = makeCluster(fmt::format(yaml_set_tcp_pool_idle_timeout, "cluster2", "9s"));
+  EXPECT_EQ(std::chrono::seconds(9), cluster2->info()->tcpPoolIdleTimeout());
+
+  auto cluster3 = makeCluster(fmt::format(yaml_set_tcp_pool_idle_timeout, "cluster3", "0s"));
+  EXPECT_EQ(absl::nullopt, cluster3->info()->tcpPoolIdleTimeout());
+}
+
 TEST_F(ClusterInfoImplTest, TestTrackTimeoutBudgetsNotSetInConfig) {
   // Check that without the flag specified, the histogram is null.
   const std::string yaml_disabled = R"EOF(